feat: remove React frontend — Leptos is now the sole frontend
Delete apps/frontend/ (855 MB — React SPA, node_modules, Vite config,
267 JSX/JS source files). Leptos serves all routes.

Changes:
- Move shared static assets (logos, icons, manifest.json, sw.js) to
  crates/kyomi-ui/public/ for Leptos to serve
- Remove apps/server/src/frontend.rs (React SPA handler with rust-embed)
- Promote leptos_frontend.rs as the sole fallback handler with proper
  ETag caching, mime types, and no-cache for manifests
- Remove frontend-builder Docker stage from both Dockerfiles
- Remove "Build frontend" steps from CI/CD release workflow (keep MCP
  chart app build which is separate)
- Remove frontend dev scripts (start-frontend.sh, restart-frontend.sh)
- Fix compilation: add hmac dep for trial chat HMAC tokens, fix
  auth.is_owner → auth.workspace.is_owner, fix queries_remaining type

The MCP chart app (apps/mcp-chart-app/) is preserved — it's a separate
Vite build that embeds into the Rust binary for the MCP server.

+++ crates/kyomi-ui/src/server_fns/trial.rs
@@ -36,7 +36,7 @@ pub struct TrialChatResponse {
     pub response: String,
     pub message_id: String,
     pub query_count: u64,
-    pub queries_remaining: i32,
+    pub queries_remaining: u64,
     pub thinking_events: Vec<serde_json::Value>,
     /// Refreshed session token (always returned for convenience).
     pub session_token: Option<String>,
